Moved bid validation into some adapter middleware. (#772)
* Moved bidd validation into some adapter middleware.

* Ran gofmt -s to fix the build.
