+++ exchange/exchange.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net/http"
 	"runtime/debug"
-	"strings"
 	"time"
 
 	"github.com/golang/glog"
@@ -18,7 +17,6 @@ import (
 	"github.com/prebid/prebid-server/openrtb_ext"
 	"github.com/prebid/prebid-server/pbsmetrics"
 	"github.com/prebid/prebid-server/prebid_cache_client"
-	"golang.org/x/text/currency"
 )
 
 // Exchange runs Auctions. Implementations must be threadsafe, and will be shared across many goroutines.
@@ -187,11 +185,6 @@ func (e *exchange) getAllBids(ctx context.Context, cleanRequests map[openrtb_ext
 			// Add in time reporting
 			elapsed := time.Since(start)
 			brw.adapterBids = bids
-			// validate bids ASAP, so we don't waste time on invalid bids.
-			err2 := brw.validateBids(request)
-			if len(err2) > 0 {
-				err = append(err, err2...)
-			}
 			// Structure to record extra tracking data generated during bidding
 			ae := new(seatResponseExtra)
 			ae.ResponseTimeMillis = int(elapsed / time.Millisecond)
@@ -398,93 +391,4 @@ func (e *exchange) makeBid(Bids []*pbsOrtbBid, adapter openrtb_ext.BidderName) (
 		}
 	}
 	return bids, errList
-}
-
-// validateBids will run some validation checks on the returned bids and excise any invalid bids
-func (brw *bidResponseWrapper) validateBids(request *openrtb.BidRequest) (err []error) {
-	// Exit early if there is nothing to do.
-	if brw.adapterBids == nil || len(brw.adapterBids.bids) == 0 {
-		return
-	}
-
-	err = make([]error, 0, len(brw.adapterBids.bids))
-
-	// By design, default currency is USD.
-	if cerr := validateCurrency(request.Cur, brw.adapterBids.currency); cerr != nil {
-		brw.adapterBids.bids = nil
-		err = append(err, cerr)
-		return
-	}
-
-	validBids := make([]*pbsOrtbBid, 0, len(brw.adapterBids.bids))
-	for _, bid := range brw.adapterBids.bids {
-		if ok, berr := validateBid(bid); ok {
-			validBids = append(validBids, bid)
-		} else {
-			err = append(err, berr)
-		}
-	}
-	if len(validBids) != len(brw.adapterBids.bids) {
-		// If all bids are valid, the two slices should be equal. Otherwise replace the list of bids with the valid bids.
-		brw.adapterBids.bids = validBids
-	}
-	return err
-}
-
-// validateCurrency will run currency validation checks and return true if it passes, false otherwise.
-func validateCurrency(requestAllowedCurrencies []string, bidCurrency string) error {
-	// Default currency is `USD` by design.
-	defaultCurrency := "USD"
-	// Make sure bid currency is a valid ISO currency code
-	if bidCurrency == "" {
-		// If bid currency is not set, then consider it's default currency.
-		bidCurrency = defaultCurrency
-	}
-	currencyUnit, cerr := currency.ParseISO(bidCurrency)
-	if cerr != nil {
-		return cerr
-	}
-	// Make sure the bid currency is allowed from bid request via `cur` field.
-	// If `cur` field array from bid request is empty, then consider it accepts the default currency.
-	currencyAllowed := false
-	if len(requestAllowedCurrencies) == 0 {
-		requestAllowedCurrencies = []string{defaultCurrency}
-	}
-	for _, allowedCurrency := range requestAllowedCurrencies {
-		if strings.ToUpper(allowedCurrency) == currencyUnit.String() {
-			currencyAllowed = true
-			break
-		}
-	}
-	if currencyAllowed == false {
-		return fmt.Errorf(
-			"Bid currency is not allowed. Was '%s', wants: ['%s']",
-			currencyUnit.String(),
-			strings.Join(requestAllowedCurrencies, "', '"),
-		)
-	}
-
-	return nil
-}
-
-// validateBid will run the supplied bid through validation checks and return true if it passes, false otherwise.
-func validateBid(bid *pbsOrtbBid) (bool, error) {
-	if bid.bid == nil {
-		return false, fmt.Errorf("Empty bid object submitted.")
-	}
-	// These are the three required fields for bids
-	if bid.bid.ID == "" {
-		return false, fmt.Errorf("Bid missing required field 'id'")
-	}
-	if bid.bid.ImpID == "" {
-		return false, fmt.Errorf("Bid \"%s\" missing required field 'impid'", bid.bid.ID)
-	}
-	if bid.bid.Price <= 0.0 {
-		return false, fmt.Errorf("Bid \"%s\" does not contain a positive 'price'", bid.bid.ID)
-	}
-	if bid.bid.CrID == "" {
-		return false, fmt.Errorf("Bid \"%s\" missing creative ID", bid.bid.ID)
-	}
-
-	return true, nil
 }
